[gejian liao] feat: parking car into parking lot of max empty ratio .

+++ src/main/java/com/example/AEPB/ParkingLot.java
@@ -4,7 +4,6 @@ package com.example.AEPB;
 import lombok.Getter;
 import org.apache.logging.log4j.util.Strings;
 
-import java.math.BigDecimal;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
@@ -60,8 +59,8 @@ public class ParkingLot {
         return maxCapacity - parkingSpace.size();
     }
 
-    public BigDecimal emptyRatio() {
-        return BigDecimal.valueOf(1 - parkingSpace.size() / maxCapacity);
+    public double emptyRatio() {
+        return 1 - (double) parkingSpace.size() / maxCapacity;
     }
 
     public boolean containCar(Car car) {
